+++ lib/screens/home.dart
@@ -1,0 +1,81 @@
+import 'package:flutter/material.dart';
+
+class Home extends StatefulWidget {
+  const Home({super.key, required this.title});
+
+  final String title;
+
+  @override
+  State<Home> createState() => _HomeState();
+}
+
+class _HomeState extends State<Home> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: Text(widget.title),
+      ),
+      drawer: const AppDrawer(),
+      body: const Center(
+        child: Column(
+          children: <Widget>[
+            Text(
+              'Herzlich Willkommen im Campus Guide!',
+              style: TextStyle(
+                fontSize: 20,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+class AppDrawer extends StatelessWidget {
+  const AppDrawer({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Drawer(
+      child: ListView(
+        padding: EdgeInsets.zero,
+        children: [
+          ListTile(
+            leading: const Icon(
+              Icons.calendar_month_outlined,
+            ),
+            title: const Text('Kalender'),
+            onTap: () {
+              Navigator.pop(context);
+              Navigator.pushNamed(context, '/calendar');
+            },
+          ),
+          ListTile(
+            leading: const Icon(
+              Icons.chat,
+            ),
+            title: const Text('Nachrichten'),
+            onTap: () {
+              Navigator.pop(context);
+              Navigator.pushNamed(context, '/news');
+            },
+          ),
+          ListTile(
+            leading: const Icon(
+              Icons.location_city,
+            ),
+            title: const Text('Locations'),
+            onTap: () {
+              Navigator.pop(context);
+              Navigator.pushNamed(context, '/locations');
+            },
+          ),
+        ],
+      ),
+    );
+  }
+}
